+++ lib/Controls/ActionRunner.js
@@ -1,5 +1,6 @@
 import CoreBase from '../Core/Base.js'
 import { CreateBankControlId, ParseControlId } from '../Shared/ControlId.js'
+import PropertySetter from './PropertySetter.js'
 
 /**
  * Class to handle execution of actions.
@@ -30,10 +31,19 @@ export default class ActionRunner extends CoreBase {
 	#timers_running = new Map()
 
 	/**
+	 * Property setter helper
+	 * @access private
+	 * @type {PropertySetter}
+	 */
+	#propertySetter
+
+	/**
 	 * @param {Registry} registry - the application core
 	 */
 	constructor(registry) {
 		super(registry, 'action-runner', 'Control/ActionRunner')
+
+		this.#propertySetter = new PropertySetter(registry)
 	}
 
 	/**
@@ -116,64 +126,9 @@ export default class ActionRunner extends CoreBase {
 			const instance = this.instance.moduleHost.getChild(action.instance)
 			if (instance) {
 				if (action.propertyId) {
-					const propertyDefinition = this.instance.definitions.getPropertyDefinition(action.instance, action.propertyId)
-					if (propertyDefinition) {
-						if (action.action !== 'set-value') throw new Error('Not implemented!')
-
-						Promise.resolve()
-							.then(async () => {
-								let instanceId = null
-
-								// If the property has instanceIds, then determine what the user has chosen and validate it
-								if (propertyDefinition.instanceIds) {
-									if (action.options.instanceUseExpression) {
-										instanceId = this.instance.variable.parseExpression(action.options.instanceExpression).value
-									} else {
-										instanceId = action.options.instanceValue
-									}
-
-									// Try a strict then loose match
-									const match =
-										propertyDefinition.instanceIds.find((inst) => inst.id === instanceId) ||
-										propertyDefinition.instanceIds.find((inst) => inst.id == instanceId)
-									if (!match) {
-										throw new Error(`Invalid instanceId: ${instanceId}`)
-									}
-									instanceId = match.id
-								}
-
-								// Parse the value
-								let value = action.options.value
-								if (action.options.valueUseExpression) {
-									if (
-										propertyDefinition.type !== 'number' &&
-										propertyDefinition.type !== 'string' &&
-										propertyDefinition.type !== 'boolean' &&
-										propertyDefinition.type !== 'dropdown'
-									)
-										throw new Error(`Cannot parse an expression to type: "${propertyDefinition.type}"`)
-
-									value = this.instance.variable.parseExpression(
-										action.options.valueExpression,
-										propertyDefinition.type !== 'dropdown' ? propertyDefinition.type : undefined
-									).value
-
-									// TODO - check value looks valid?
-								} else if (propertyDefinition.type === 'string') {
-									// if a string, we need to parse variables in the normal value
-									value = this.instance.variable.parseVariables(action.options.value)
-								}
-
-								await instance.propertySet(action.propertyId, instanceId, value, extras)
-							})
-							.catch((e) => {
-								this.logger.warn(`Error setting property for ${instance.connectionId}: ${e.message ?? e}`)
-							})
-					} else {
-						this.logger.info(
-							`Skipping unknown property ${action.propertyId} for ${instance.connectionId}: ${e.message ?? e}`
-						)
-					}
+					this.#propertySetter.runAction(instance, action, extras).catch((e) => {
+						this.logger.warn(`Error setting property for ${instance.connectionId}: ${e.message ?? e}`)
+					})
 				} else {
 					instance.actionRun(action, extras).catch((e) => {
 						this.logger.silly(`Error executing action for ${instance.connectionId}: ${e.message ?? e}`)
